requirements.txt
numpy
opencv_python==4.10.0.82
Pillow==10.3.0
PyYAML==6.0.1
streamlit==1.32.0
ultralytics==8.0.230
splitfolders